@@ -8,29 +8,31 @@
   "license": "BSD-3-Clause",
   "main": "index.js",
   "scripts": {
-    "zapier-build": "tsc",
-    "zapier-dev": "tsc --watch",
+    "zapier-clean": "rimraf ./lib || rimraf ./build",
+    "zapier-build": "npm run zapier-clean && tsc",
+    "zapier-dev": "npm run zapier-clean && tsc --watch",
+    "zapier-validate": "npm run zapier-build && zapier validate",
     "zapier-push": "npm run zapier-build && zapier push",
     "prepare": "npm run zapier-build",
     "pretest": "npm run zapier-build",
-    "test": "./node_modules/.bin/mocha --compilers ts:ts-node/register ./src/test/*.ts"
+    "test": "mocha --compilers ts:ts-node/register ./src/test/*.ts"
   },
   "engines": {
     "node": ">=8.10.0",
     "npm": ">=5.6.0"
   },
   "dependencies": {
-    "@types/multiparty": "0.0.32",
-    "multipart-stream": "^2.0.1",
+    "form-data": "^3.0.0",
     "zapier-platform-core": "9.0.0"
   },
   "devDependencies": {
     "@types/mocha": "5.2.0",
-    "@types/node": "10.1.2",
+    "@types/node": "10.17.9",
     "mocha": "5.1.0",
+    "rimraf": "3.0.0",
     "should": "13.2.1",
-    "ts-node": "^8.3.0",
-    "typescript": "3.3"
+    "ts-node": "^8.6.2",
+    "typescript": "3.7.3"
   },
   "private": true
 }
