@@ -2,13 +2,13 @@ export const Routes = {
     GetAccount: "/api/integration/v1/account",
     CreateProject: "/api/integration/v1/project/create",
     GetProject: "/api/integration/v1/project",
-    GetProjects: "/api/integration/v1/project/list",
-    UploadDocument: "/api/integration/v1/project/document",
+    GetProjectList: "/api/integration/v1/project/list",
+    UploadProjectDocument: "/api/integration/v1/project/document",
     UpdateDocument: "/api/integration/v1/document/update",
-    CreateExport: "/api/integration/v1/document/export",
-    GetVendors: "/api/integration/v1/directory?type=vendor",
-    GetCountries: "/api/integration/v1/directory?type=country",
-    GetClients: "/api/integration/v1/directory?type=client",
+    CreateDocumentExport: "/api/integration/v1/document/export",
+    GetDirectoryVendors: "/api/integration/v1/directory?type=vendor",
+    GetDirectoryCountries: "/api/integration/v1/directory?type=country",
+    GetDirectoryClients: "/api/integration/v1/directory?type=client",
     GetClient: "/api/integration/v2/client",
     CreateJob: "/api/integration/v2/invoice/job"
 };
